+++ CURLHandleSource/CURLSocketRegistration.m
@@ -44,7 +44,7 @@
     [super dealloc];
 }
 
-- (void)updateSourcesForSocket:(int)socket mode:(NSInteger)mode multi:(CURLMultiHandle*)multi
+- (void)updateSourcesForSocket:(int)socket mode:(int)mode multi:(CURLMultiHandle*)multi
 {
     // We call back to the multi to do the actual work - this class really just exists as
     // a place to group together the reader and writer sources corresponding to a socket.
